requirements.txt
streamlit==1.37.0
numpy==1.24.3
pandas==2.0.3
Pillow==9.0.0
joblib==1.3.2
seaborn==0.13.2
matplotlib==3.7.5